Added support for Docker to CenOS 7 tree

+++ centos-7/scripts/base.sh
@@ -8,18 +8,22 @@ yum -y install deltarpm ntpdate
 
 ntpdate -s time.nist.gov
 
-yum -y install binutils fuse-libs gcc gcc-c++ kernel-devel-`uname -r` make perl yum-utils
+yum -y install binutils fuse-libs gcc gcc-c++ make perl yum-utils
+
+if [ ! -f "/.dockerinit" ]; then
+  yum -y install kernel-devel-`uname -r`
+fi
 
 yum -y upgrade
 
 # Setup MOTD
 cat > /etc/motd << 'EOF'
-         ______           __  ____  _____    _____   ____                
-        / ____/__  ____  / /_/ __ \/ ___/   /__  /  / __ )____ _________ 
+         ______           __  ____  _____    _____   ____
+        / ____/__  ____  / /_/ __ \/ ___/   /__  /  / __ )____ _________
        / /   / _ \/ __ \/ __/ / / /\__ \      / /  / __  / __ `/ ___/ _ \
       / /___/  __/ / / / /_/ /_/ /___/ /     / /  / /_/ / /_/ (__  )  __/
-      \____/\___/_/ /_/\__/\____//____/     /_/  /_____/\__,_/____/\___/ 
-                                                                         
+      \____/\___/_/ /_/\__/\____//____/     /_/  /_____/\__,_/____/\___/
+
              __             ______                ____
             / /_  __  __   / ____/__  ____  ___  / __ )___  ____ _____
            / __ \/ / / /  / / __/ _ \/ __ \/ _ \/ __  / _ \/ __ `/ __ \
@@ -33,5 +37,7 @@ echo 'Testing the MOTD...'
 echo
 cat /etc/motd
 
-reboot
-sleep 60
+if [ ! -f "/.dockerinit" ]; then
+  reboot
+  sleep 60
+fi
